Sprint 1 - User story 2 - Refactored reset password routes
Created a new controller to request a password reset and updated other client
and server code accordingly. This will result in less merge conflicts when merging develop to
feature/002/forgotPasswordStep
